moved null replacement loader to utils
